deps!: update multiformats to 10.x.x
Updates the multiformats dep to the new version. This module exposes
the `CID` class as part of the public API so it's safest if this goes
out as a major.

BREAKING CHANGE: the `CID` class used as properties of the interfaces from this module now comes from v10.x.x of the multiformats module