deps!: update multiformats to v13
v13 of multiformats was release which is a [port to TypeScript](https://github.com/multiformats/js-multiformats/releases/tag/v13.0.0).

There are no API changes so the upgrade should be relatively
painless but it's been released as a major for safety.

BREAKING CHANGE: multiformats have been upgraded to v13 - the API should remain identical